# api-quiz/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/quiz/categories": {
      "get": {
        "operationId": "indexCategories",
        "summary": "",
        "description": "Index Categories",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Category name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": false,
            "in": "query",
            "description": "Category slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "Category code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoriesResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "categories"
        ]
      }
    },
    "/api/quiz/admin/categories": {
      "get": {
        "operationId": "indexCategories",
        "summary": "",
        "description": "Index Categories",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Category name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": false,
            "in": "query",
            "description": "Category slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "Category code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoriesResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin.categories"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "createCategory",
        "summary": "",
        "description": "Create Category",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin.categories"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Quiz API",
    "description": "Swagger specification for api-quiz | [swagger.json](swagger.json)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "message": {
            "type": "string",
            "example": "Error message",
            "default": "Internal Server Error"
          },
          "name": {
            "type": "string",
            "example": "Error"
          }
        }
      },
...
      "CategoriesResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Đánh giá tư duy"
          },
          "slug": {
            "type": "string",
            "description": "Slug",
            "example": "danh-gia-tu-duy"
          },
          "code": {
            "type": "string",
            "description": "Code",
            "example": "VN1111"
          },
          "imageUrl": {
            "type": "string",
            "description": "Image Url",
            "example": "https://www.studyphim.vn/system/movies/486/thumbnails/medium/Untitled.png"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateCategoryInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Đánh giá tư duy"
          },
          "slug": {
            "type": "string",
            "description": "Slug",
            "example": "danh-gia-tu-duy"
          },
          "code": {
            "type": "string",
            "description": "Code",
            "example": "VN1111"
          },
          "imageUrl": {
            "type": "string",
            "description": "Image Url",
            "example": "https://www.studyphim.vn/system/movies/486/thumbnails/medium/Untitled.png"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}